OboeTester: refactor tap-to-tone
Separate MIDI from the audio listener.
Prepare for external tap-to-tone class.
